pyproject.toml
[project]
name = "private-universities-sms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django-cotton>=1.5.1",
    "django==5.1.0",
    "django-filter>=24.3",
    "django-extensions>=3.2.3",
    "djangoql>=0.18.1",
    "django-import-export[xlsx]>=4.3.3",
    "heroicons[django]>=2.9.0",
    "django-htmx>=1.21.0",
    "mysqlclient>=2.2.6",
    "python-dotenv>=1.0.1",
    "django-widget-tweaks>=1.5.0",
    "django-braces>=1.16.0",
    "django-content-settings>=0.29.1",
]

[dependency-groups]
dev = [
    "coverage>=7.6.9",
    "django-silk>=5.3.1",
    "icecream>=2.1.3",
    "selectolax>=0.3.27",
]
